refinement 1

+++ Book.cpp
@@ -45,7 +45,7 @@ void Book::printBook(){
 
 float Book::getPopRate(){
     if(this->total_num_of_likes == 0) return 0.0;
-    else return (this->author_average_rating + (static_cast<float>(this->numerator)/this->total_num_of_likes))/10;
+    else return (this->author_average_rating + this->numerator/this->total_num_of_likes)/10;
 }
 
 void Book::updateRates(int rate, int number_of_like){
